+++ src/CodeWF.AvaloniaControls.ProDataGrid.Themes/CodeWFProDataGridTheme.axaml.cs
@@ -1,0 +1,7 @@
+using Avalonia.Styling;
+
+namespace CodeWF.AvaloniaControls.ProDataGrid.Themes;
+
+public class CodeWFProDataGridTheme : Styles
+{
+}