+++ src/CodeWF.AvaloniaControls.ProDataGrid.Themes/Properties/AssemblyInfo.cs
@@ -1,0 +1,4 @@
+using Avalonia.Metadata;
+
+[assembly: XmlnsPrefix("https://codewf.com", "codewf")]
+[assembly: XmlnsDefinition("https://codewf.com", "CodeWF.AvaloniaControls.ProDataGrid.Themes")]