+++ src/CodeWF.AvaloniaControls.ProDataGridDemo/OwnerTheme.axaml.cs
@@ -1,0 +1,7 @@
+using Avalonia.Styling;
+
+namespace CodeWF.AvaloniaControls.ProDataGridDemo;
+
+public class OwnerTheme : Styles
+{
+}