+++ src/CodeWF.AvaloniaControls.ProDataGridDemo/ViewModels/ViewModelBase.cs
@@ -1,0 +1,7 @@
+using ReactiveUI;
+
+namespace CodeWF.AvaloniaControls.ProDataGridDemo.ViewModels;
+
+public class ViewModelBase : ReactiveObject
+{
+}