+++ src/CodeWF.AvaloniaControls.ProDataGridDemo/Views/DocumentWorkspaceView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia.Controls;
+
+namespace CodeWF.AvaloniaControls.ProDataGridDemo.Views;
+
+public partial class DocumentWorkspaceView : UserControl
+{
+    public DocumentWorkspaceView()
+    {
+        InitializeComponent();
+    }
+}